requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-cors-headers==3.1.0
django-environ==0.4.5
djangorestframework==3.10.3
ff-espn-api==1.1.2
get==2019.4.13
idna==2.8
nose==1.3.7
numpy==1.16.4
post==2019.4.13
public==2019.4.13
pytz==2019.2
query-string==2019.4.13
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.5
